Proper errors from memory ops
