+++ test/fail/turnt.toml
@@ -1,2 +1,2 @@
 command = "bril2json < {base}.bril | brili"
-return_code = 1
+return_code = 2
